North Star: Add test images.
These are "synthetic" test images, in that the .nsi's were created
by hand to include track and sector information.  Then they were
written to real disks using FluxEngine and read back.

These are useful as they help ensure that the North Star sector
counting is accurate, as there is no track or sector information
encoded in the on-disk sector itself.